+++ src/main/java/es/tid/graphlib/als/IntDoubleArrayHashMapTextVertexOutputFormat.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import org.apache.giraph.graph.Vertex;
 import org.apache.giraph.io.formats.TextVertexOutputFormat;
 import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.TaskAttemptContext;
@@ -19,7 +20,7 @@ import es.tid.graphlib.utils.DoubleArrayListHashMapWritable;
  */
 public class IntDoubleArrayHashMapTextVertexOutputFormat extends
   TextVertexOutputFormat<IntWritable, DoubleArrayListHashMapWritable,
-  IntWritable> {
+  DoubleWritable> {
 
   /** Specify the output delimiter */
   public static final String LINE_TOKENIZE_VALUE = "output.delimiter";
@@ -55,7 +56,7 @@ public class IntDoubleArrayHashMapTextVertexOutputFormat extends
 
     @Override
     protected Text convertVertexToLine
-    (Vertex<IntWritable, DoubleArrayListHashMapWritable, IntWritable, ?>
+    (Vertex<IntWritable, DoubleArrayListHashMapWritable, DoubleWritable, ?>
           vertex)
       throws IOException {
